+++ app/src/main/java/com/example/allen/brofinder/activity/MapsActivity.java
@@ -33,8 +33,8 @@ public class MapsActivity extends FragmentActivity implements
     private static final String DESTINATION_LAT_PARAM = "latitude";
     private static final String DESTINATION_LON_PARAM = "longitude";
 
-    private float destinationLat;
-    private float destinationLon;
+    private double destinationLat;
+    private double destinationLon;
 
     @Override
     protected void onCreate(Bundle savedInstanceState) {
@@ -48,8 +48,8 @@ public class MapsActivity extends FragmentActivity implements
                 .addOnConnectionFailedListener(this)
                 .build();
 
-        destinationLat = Float.parseFloat(getIntent().getExtras().getString(DESTINATION_LAT_PARAM));
-        destinationLon = Float.parseFloat(getIntent().getExtras().getString(DESTINATION_LON_PARAM));
+        destinationLat = Double.parseDouble(getIntent().getExtras().getString(DESTINATION_LAT_PARAM));
+        destinationLon = Double.parseDouble(getIntent().getExtras().getString(DESTINATION_LON_PARAM));
         destinationLocation = new Location("");
         destinationLocation.setLatitude(destinationLat);
         destinationLocation.setLongitude(destinationLon);
